Full CRUD for questions e2e tested
Question types:
- Multiple-choice
- Single-choice
- True-or-false

Creation of a question is the trickiest, as the user might pass in a
very big context, unsuitable for the llm to ingest. For this reason
the user submitted prompt is chunked and stored in chunked in the db
and in an inmemory cache of the go server (no eviction policy yet).
Later this data can be used for RAG, and for training data for a new
model.